requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-huggingface
langchain-pinecone
langchain-community
langgraph
pinecone-client
tavily-python
pypdf
sentence-transformers
python-dotenv
python-multipart
